Replace pkg_resources with importlib.resources.
Asset overrides now resolve through Pyramid's own APIs via explicit
IPackageOverrides registry lookup, removing the setuptools runtime
dependency and the OverrideProvider loader hook.
