+++ pyproject.toml
@@ -23,7 +23,6 @@ dependencies = [
     "hupper >= 1.5",  # ignore_files support
     "plaster",
     "plaster_pastedeploy",
-    "setuptools < 82",  # require pkg_resources
     "translationstring >= 0.4",  # py3 compat
     "venusian >= 1.0",  # ``ignore``
     "webob >= 1.8.3",  # Accept.parse_offer
