+++ tests/pkgs/assets/dir/fizz.txt
@@ -1,0 +1,1 @@
+fizz
