+++ tests/pkgs/assets/foo.txt
@@ -1,0 +1,1 @@
+foo
